+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## v0.19.2-dev
+
+- Bug fixes
+    - Short code maps are now stored in the correct file encoding (`UTF-8`) to allow parsing short code maps with characters like umlauts ([#15](https://github.com/elixytics/ua_inspector/issues/15))
+
 ## v0.19.1 (2019-01-05)
 
 - Enhancements